requirements.txt
i2cdev
spidev
smbus
pyserial

